pyproject.toml
[tool.poetry]
name = "airf-oper"
version = "0.1.0"
description = ""
authors = ["Chernyshev-Pridvorov"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
apache-airflow = "^2.9.3"
apache-airflow-providers-http = "4.12.0"
pydantic-settings = "^2.2.1"
apache-airflow-providers-postgres = "^5.11.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
